2_compute/2b_Cloud Functions/scripts/requirements.txt
# Function dependencies, for example:
# package>=version
tifffile
google-cloud-storage